+++ packages/client/src/adapters/fetch/orpc-link.ts
@@ -1,6 +1,6 @@
-import type { ClientContext, HTTPMethod } from '@orpc/contract'
+import type { HTTPMethod } from '@orpc/contract'
 import type { Value } from '@orpc/shared'
-import type { ClientLink, ClientOptionsOut } from '../../types'
+import type { ClientContext, ClientLink, ClientOptionsOut } from '../../types'
 import type { FetchWithContext } from './types'
 import { ORPCError } from '@orpc/contract'
 import { isAsyncIteratorObject, type StandardBody } from '@orpc/server-standard'
