+++ packages/server/src/index.ts
@@ -19,7 +19,7 @@ export * from './router-accessible-lazy'
 export * from './router-client'
 export * from './utils'
 
-export { isDefinedError, ORPCError } from '@orpc/client'
-export { eventIterator, safe, type, ValidationError } from '@orpc/contract'
+export { isDefinedError, ORPCError, safe } from '@orpc/client'
+export { eventIterator, type, ValidationError } from '@orpc/contract'
 export { getEventMeta, withEventMeta } from '@orpc/server-standard'
 export { onError, onFinish, onStart, onSuccess } from '@orpc/shared'
