+++ packages/server/src/procedure-client.ts
@@ -1,10 +1,11 @@
+import type { Client, ClientContext } from '@orpc/client'
 import type { Interceptor, MaybeOptionalOptions, Value } from '@orpc/shared'
 import type { Context } from './context'
 import type { Lazyable } from './lazy'
 import type { MiddlewareNextFn } from './middleware'
 import type { AnyProcedure, Procedure, ProcedureHandlerOptions } from './procedure'
 import { ORPCError } from '@orpc/client'
-import { type Client, type ClientContext, type ErrorFromErrorMap, type ErrorMap, type Meta, type Schema, type SchemaInput, type SchemaOutput, ValidationError } from '@orpc/contract'
+import { type ErrorFromErrorMap, type ErrorMap, type Meta, type Schema, type SchemaInput, type SchemaOutput, ValidationError } from '@orpc/contract'
 import { intercept, toError, value } from '@orpc/shared'
 import { createORPCErrorConstructorMap, type ORPCErrorConstructorMap, validateORPCError } from './error'
 import { unlazy } from './lazy'
